Update BT_CTLR_SDC_RX_PACKET_COUNT and BT_CTLR_SDC_TX_PACKET_COUNT description
Update the description of BT_CTLR_SDC_RX_PACKET_COUNT and BT_CTLR_SDC_TX_PACKET_COUNT to make it more clear on that the packet count should only be increased if the application is not able to provide/read data to/from the controller. Or set the count  to lower value if the reduced throughput is accepted. 
Signed-off-by: Hung Bui <hung.bui@nordicsemi.no>